+++ server/camcops_server/cc_modules/cc_html.py
@@ -434,8 +434,7 @@ def pmid(x: int) -> str:
     Returns hyperlinked text to a PubMed ID (PMID).
 
     Args:
-        p:
-            The integer PMID.
+        x: The integer PMID.
 
     Returns:
         Hyperlinked text, as raw HTML.
@@ -449,11 +448,22 @@ def doi(x: str) -> str:
     Returns hyperlinked text to a digital object identifier (DOI).
 
     Args:
-        p:
-            The integer PMID.
+        x: The DOI.
 
     Returns:
         Hyperlinked text, as raw HTML.
 
     """
     return f'<a href="https://doi.org/{x}">doi:{x}</a>'
+
+
+def a_href(url: str, text: str = None) -> str:
+    """
+    Returns text hyperlinked to a URL; by default, the text is the URL itself.
+
+    Args:
+        url: the raw URL
+        text: text to be shown
+    """
+    text = text or url
+    return f'<a href="{url}">{text}</a>'